feat(ObservableObjectMap): Added methods to remove and retrieve MapChangeListeners
This commits adds the removeListener() method of the ObservableObjectMap class to remove a specified MapChangeListener from the list of registered listeners. The getListeners() method has also been added to retrieve an unmodifiable list of all registered MapChangeListeners. The method utilizes the Collections.unmodifiableList() method to ensure that the list cannot be modified externally. This update enhances the flexibility and usability of the ObservableObjectMap class.